requirements.txt
dnspython==2.1.0
termcolor==1.1.0
emailprotectionslib==0.